tests,file_ops,web_fetch: retire two stale assertions and drop dead bindings
Four substantive changes plus the clippy micro-fixes that came with them.

`web_fetch`: `uses_fetch_provider_before_builtin` asserted the opposite of
the shipped behaviour. BT-D-R4-22 deliberately skips configured fetch
providers — the validated DNS pin cannot be threaded into a provider's own
HTTP client, so consulting one would reopen the rebinding window between
validate and connect. The test still expected provider content, and it
reached for example.com to get it. Rewritten as a spy that must NOT be
called, against a dead local port so the built-in fetch fails fast with no
external network. A deliberate behaviour change whose tests were not
updated in the same batch is indistinguishable from a regression.

`file_search/walk.rs`: `floor_skipped_dirs` was measuring the environment.
The walker honours .gitignore files in PARENTS of the root, and some
machines ship a $TMPDIR/.gitignore naming node_modules/ — which skips the
directory before the floor filter ever sees it, leaving the counter at 0
for a reason the test did not arrange. A local `!node_modules/` negation
keeps the directory visible to the gitignore layer, so the skip counted
below is the floor filter's own work.

`external_link.rs`: the internal-URL test asserted `data:text/html,hello`
is internal, which `data_urls_are_not_internal` two tests down forbids.
Removed, with a pointer to the test that owns the rule.

`gateway/handlers/pty.rs`: bind `terminal` through its named type so
scripts/config_wiring_audit.py can see that [policies.terminal] has a
consumer outside src/config/. Field inference made the consumer invisible
to the audit.

The rest is clippy: three `let _ = deleted` bindings in execute_delete
(and a `(count, ())` destructure that says what the unit actually is), a
needless_return in focus_window's macOS arm, two unnecessary `as u64`
casts, and `unwrap_or_else(Owner::new)` -> `unwrap_or_default()`.

+++ interfaces/webchat/src/platform/wide/views/settings/embedding_providers/reembed_card.rs
@@ -84,8 +84,8 @@ pub(super) fn ReembedMigrationCard() -> impl IntoView {
                             set_result_errors.set(error_list.into_iter().take(5).collect());
 
                             set_result_counts.set(Some((
-                                facts_updated as u64,
-                                facts_total as u64,
+                                facts_updated,
+                                facts_total,
                                 errors as u64,
                             )));
                         }
